+++ docs/src/pages/api/APIDeclaration/configure.tsx
@@ -9,7 +9,7 @@ import {
   Parameter,
   Parameters,
   Purpose,
-  Usage,
+  Usage
 } from '../../../components/api-components'
 import { Members, PropertyMember } from '../../../components/members-table'
 import { Link } from 'gatsby'
@@ -49,15 +49,15 @@ export default function Configure() {
               <Parameter name="url" dataType="string" />
               <Parameter name="query" dataType="object" />
               <Parameter name="body" dataType="object" />
-              <Parameter name="params" dataType="object" availableFrom='1.2.0'>
+              <Parameter name="params" dataType="object" availableFrom="1.2.0">
                 Do note that params are already baked into the URL, there is no need for the function to do that.
               </Parameter>
               <Parameter name="route" dataType="object/function">
                 This is the object/function for route being invoked. For most uses this should be completely unnecessary, but this can be
-                used to allow for special behavior for particular routes. If options have been defined on the route, you can access them from
-                route.options.
+                used to allow for special behavior for particular routes. If options have been defined on the route, you can access them
+                from route.options.
               </Parameter>
-              <Parameter name='requestOptions' dataType='varies'>
+              <Parameter name="requestOptions" dataType="varies">
                 If request options are used, then they are passed as this parameter.
               </Parameter>
             </Parameters>
@@ -79,14 +79,35 @@ export default function Configure() {
             Express router, on which the declared APIs will be attached. The main advantage of using a router over app is that APIs can be
             added after the whole application has been configured, assuming no middleware is added to the router itself after the routes.
           </PropertyMember>
-          <PropertyMember name="routerAt" dataType="string" availableFrom='1.4.0'>
-            <p>It often makes sense to declare APIs on top a common base URL. In practice though you might have a router
-              set up for a specific path, for example to add common middleware.</p>
+          <PropertyMember name="routerAt" dataType="string" availableFrom="1.4.0">
             <p>
-              By setting routerAt, you explicitly indicate that the provided router will be at the given path. All routes
-              on an API declaration with routerAt MUST have the routerAt as the start of their path.
+              It often makes sense to declare APIs on top a common base URL. In practice though you might have a router set up for a
+              specific path, for example to add common middleware.
             </p>
-
+            <p>
+              By setting routerAt, you explicitly indicate that the provided router will be at the given path. All routes on an API
+              declaration with routerAt MUST have the routerAt as the start of their path.
+            </p>
+          </PropertyMember>
+          <PropertyMember name="inherentMiddleware" dataType="Function[]" availableFrom="1.5.0">
+            An array of papupata middleware applied to all routes implemented in this API declaration, that is, middleware inherent to all
+            routes of this declaration. Only relevant for implementing, not for clients.
+          </PropertyMember>
+          <PropertyMember name="autoImplementAllAPIs" dataType="boolean" availableFrom="1.5.0">
+            <p>
+              If set to true, all declared but unimplemented routes are automatically set up to return 501 not implemented as a response to
+              requests. If there is an app or router present at the time of declaration, this happens immediately, otherwise it happens as
+              soon as an app or router is configured.
+            </p>
+            <p>
+              This two major effect; for one, attempts to call unimplemented papupata APIs do not end up going to other route handlers or
+              middleware (although with <Link to="/api/skipHandlingRoute">skipHandlingRoute</Link> this behavior can be restored as needed).
+              Also, routes are implemented in the order they were defined in, not in the order they were implemented in. Usually this does
+              not make a big difference, but if you have two handlers that can match the same path, say, /api/:var and /api/fixed, the only
+              way requests ever reached fixed was if it was implemented first. With this variable set to true, it has to be declared first,
+              but the implementation order is irrelevant.
+            </p>
+            <p>It is likely that from version 2.0 onwards the default value for this setting becomes true.</p>
           </PropertyMember>
         </Members>
         <div>*1: For invoking APIs or calling the getURL method on them</div>